requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
stripe==7.11.0
razorpay==1.4.1
django-storages==1.14.2
boto3==1.34.34
django-filter==23.5
django-rest-swagger==2.2.0
gunicorn==21.2.0
whitenoise==6.6.0